OPS-127 fix: department 생성 오류 수정

+++ src/main/java/com/sprint/example/sb01part2hrbankteam10/dto/EmployeeDistributionDto.java
@@ -10,6 +10,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class EmployeeDistributionDto {
   private String groupKey;
-  private Integer count;
+  private Long count;
   private Double percentage;
+
+//  public EmployeeDistributionDto(String groupKey, Long count) {
+//    this.groupKey = groupKey;
+//    this.count = count;
+//  }
 }
